Split dmf2mod core into separate files
