+++ include/all_modules.h
@@ -1,0 +1,14 @@
+/*
+    all_modules.h
+    Written by Dalton Messmer <messmer.dalton@gmail.com>.
+
+    Includes the headers for all supported modules.
+*/
+
+#pragma once
+
+// Include all supported modules here. Should just be the contents of the include/modules folder.
+// TODO: Find a way to automate this
+
+#include "modules/dmf.h"
+#include "modules/mod.h"
