+++ include/dmf2mod.h
@@ -1,0 +1,14 @@
+/*
+    dmf2mod.h
+    Written by Dalton Messmer <messmer.dalton@gmail.com>.
+
+    Include this header for access to dmf2mod core functionality.
+*/
+
+#pragma once
+
+// Include module and conversion options definitions, registrar, etc.
+#include "core/module.h"
+
+// Include headers for all the supported modules
+#include "all_modules.h"
